QArchSearch/requirements.txt
qiskit>=0.45
networkx>=3.1
numpy>=1.24
scipy>=1.11
pyyaml>=6.0
pytest>=7.4

